feat: Add unit test for generatespeedtable
