+++ traffic_updater/go/osrm_traffic_updater/testdata/target.csv
@@ -1,0 +1,5 @@
+84760891102,19496208102,81
+84762609102,244183320001101,87
+244183320001101,84762607102,87
+84760849102,84760850102,47
+84760846102,84760858102,59
